fix(tests): fold the store-isolation guard into the incarnations mirror
CI was red on both matrix legs and I could not reproduce it locally. The
reason turned out to be structural rather than flaky: `test_audit_all_clean`
is SKIPPED on a developer host and RUNS in the CI SIF, so no local run of
`tests/develop/` could ever have caught this.

The failure, once the Actions log was readable:

    ERRO: [E] [PS-204 §2 orphan-test-file]
      tests/scitex_agent_container/test_store_isolation_guard.py

A standalone `tests/scitex_agent_container/test_*.py` has no src
counterpart, and the ecosystem audit rejects it. Nothing about the store
guard itself — the 357 StoreTargetError lines in the log are the guard
working exactly as designed, and are noise from best-effort writers, not
failures. 17,211 tests passed; one gate failed on the file's LOCATION.

WHY NOT tests/develop/, WHICH IS WHERE SUCH GATES USUALLY GO
============================================================
It would satisfy the audit — those files have no src mirror and pass. But
`tests/develop/` is OUTSIDE the conftest whose guard these tests verify,
so the guard would not be armed there and every assertion would pass by
not being covered. A test that passes because the thing it checks is
absent is the failure mode this whole PR exists to remove.

So they move to `_state/test_state_db_incarnations.py`: the guard applies
(same conftest scope), the `pg_schema` fixture is available, the file
mirrors a real src module, and it is the table the incident actually
polluted.

VERIFIED WITH THE INSTRUMENT, NOT BY INFERENCE
==============================================
Removing the orphan file makes the cause absent, which is not the same as
the gate passing. Ran the auditor CI runs:

    python -m scitex_dev ecosystem audit-all scitex-agent-container
    -> summary: 0 unmasked error(s), 0 masked by skip-rules; exit 0

20 tests pass in the merged file; ruff clean on CI's exact invocation.
